+++ CHANGELOG.md
@@ -18,7 +18,7 @@ Senior Web Developer
 - **(enh)** Add option to disable asset publication in debug mode #229 (suralc)
 - **(fix)** Fixed the CSS-including behavior of TbHtml5Editor #290 #311 (hijarian)
 - **(fix)** Now assets for TbDateRangePicker are being registered with default settings instead of hard-coded POS_HEAD #266 #297 (hijarian)
-
+- **(fix)** Added correct handling of the composite primary keys in the `TbEditableField` #287 (abeel)
 
 ### YiiBooster version 1.0.5